requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.51.0
PyPDF2==3.0.1
pdf2image==1.16.3
moviepy==1.0.3
Pillow==10.0.1
gtts==2.4.0
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3